send strategy:  Increase SO_SNDBUF and detect losses.
Increase SO_SNDBUF to avoid frequent buffer overruns leading to packet
loss (since we're using non-blocking operation).  Also, report lost
packets on the console.

As a note, the practically most relevant SNDBUF change was for the
inside_fd_other that we hand over to BRT, which does not set its own
SNDBUF size and is written with blocking operation in mind.